Split multiplexer tests
