+++ test/CMakeLists.txt
@@ -42,6 +42,7 @@ make_test(test_log_to_file)
 make_test(test_conn_logging)
 make_test(test_reader_fsm)
 make_test(test_hello_utils)
+make_test(test_multiplexer)
 
 # Tests that require a real Redis server
 make_test(test_conn_quit)
