+++ test/Jamfile
@@ -58,6 +58,7 @@ local tests =
     test_conn_logging
     test_reader_fsm
     test_hello_utils
+    test_multiplexer
 ;
 
 # Build and run the tests
